#13 fix on s2_cloud
