+++ eo_forge/io/SentinelLoaders.py
@@ -5,9 +5,7 @@ Sentinel loaders module
     :toctree: ../generated/
 
     Sentinel2Loader
-    s2_cloud_preproc
 """
-import geopandas as gpd
 import numpy as np
 import os
 import rasterio as rio
@@ -22,45 +20,12 @@ from eo_forge.utils.sentinel import (
     SENTINEL2_BANDS_RESOLUTION,
     SENTINEL2_SUPPORTED_RESOLUTIONS,
     calibrate_sentinel2,
-    s2_metadata,
     calibrate_s2_scl,
+    s2_cloud_preproc,
+    s2_metadata,
 )
-from eo_forge.utils.utils import walk_dir_files
-
 
 ######################################################################
-
-
-def s2_cloud_preproc(base_dir, dump_file=None):
-    """
-    Read cloud mask file as geodataframe and write to disk (if necessary)
-    :param dump_file: file to be written (if None, just return the
-    geodataframe)
-    """
-
-    _, _, g = walk_dir_files(base_dir, cases=["MSK_CLOUDS_B00.gml"])
-
-    if "MSK_CLOUDS_B00.gml" in g:
-        mask_cloud_file_ = g["MSK_CLOUDS_B00.gml"][0]
-    else:
-        mask_cloud_file_ = None
-
-    gpd_ = None
-
-    if mask_cloud_file_ is None:
-        gpd_ = None
-        return gpd_
-    else:
-        try:
-            gpd_ = gpd.read_file(mask_cloud_file_)
-            if dump_file is None:
-                pass
-            else:
-                gpd_.to_file(dump_file)
-            return gpd_
-        except:  # noqa
-            print(f"FAILED to read/dump file: {mask_cloud_file_}")
-            return None
 
 
 class Sentinel2Loader(BaseGenericLoader):
