+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = eo-forge
-version = 0.3.0
+version = 0.3.1
 author = Felix Rojo Lapalma, Andrés Pérez Hortal
 description = Python tools for remote sensing data processing.
 long_description = file: README.rst
